blog/requirements.txt
appdirs==1.4.0
bootstrap-admin==0.3.7.1
Django==1.10
django-pagedown==0.1.1
Markdown==2.6.7
markdown2==2.3.2
meld3==1.0.2
MySQL-python==1.2.5
packaging==16.8
pyparsing==2.1.10
six==1.10.0
supervisor==3.3.1
uWSGI==2.0.14
